idPool: Remove useless uint casts
stringtable supports size_t under the hood, there is no reason to restrict it to uint.

+++ src/dmd/identifier.h
@@ -40,7 +40,7 @@ public:
     static StringTable fullPathStringTable;
     static Identifier *generateId(const char *prefix);
     static Identifier *generateId(const char *prefix, size_t i);
-    static Identifier *idPool(const char *s, unsigned len);
+    static Identifier *idPool(const char *s, size_t len);
 
     static inline Identifier *idPool(const char *s)
     {
